Backport a cast to silence a type mismatch warning
git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_1_2@7810 dcc99617-32d9-48b4-a31d-7c20da2025e4

+++ Open-ILS/src/apachemods/mod_xmlent.c
@@ -342,7 +342,7 @@ static int xmlEntHandler( ap_filter_t *f, apr_bucket_brigade *brigade ) {
 				/* log and die on XML errors */
 				ap_log_rerror( APLOG_MARK, APLOG_ERR, 0, f->r, "XMLENT XML Parse Error: %s at line %d\n",
 					XML_ErrorString(XML_GetErrorCode(ctx->parser)), 
-					XML_GetCurrentLineNumber(ctx->parser));
+					(int) XML_GetCurrentLineNumber(ctx->parser));
 
 				XML_ParserFree(parser);
 				parser = NULL;
